ILI9341 driver moved to devices_lib
